+++ src/main/java/com/netlify/hparcells/chaos/events/DiamondEvent.java
@@ -1,0 +1,21 @@
+package com.netlify.hparcells.chaos.events;
+
+import com.netlify.hparcells.chaos.ChaosEvent;
+import org.bukkit.Bukkit;
+import org.bukkit.Material;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
+
+public class DiamondEvent extends ChaosEvent {
+    public void onEnable(String username) {
+        Player target = (Bukkit.getServer().getPlayer(username));
+
+        if(target != null) {
+            target.getInventory().addItem(new ItemStack(Material.DIAMOND));
+        }
+    }
+
+    public String getName() {
+        return "Diamond";
+    }
+}