+++ src/main/java/com/netlify/hparcells/chaos/events/NiceXpEvent.java
@@ -1,0 +1,19 @@
+package com.netlify.hparcells.chaos.events;
+
+import com.netlify.hparcells.chaos.ChaosEvent;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
+
+public class NiceXpEvent extends ChaosEvent {
+    public void onEnable(String username) {
+        Player target = (Bukkit.getServer().getPlayer(username));
+
+        if(target != null) {
+            target.setLevel(69);
+        }
+    }
+
+    public String getName() {
+        return "Nice XP";
+    }
+}